add modal

+++ src/app/list-defunt/list-defunt.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Defunt } from '../models/Defunt';
 import { DefuntSearch } from '../models/DefuntSearch';
 import { DefuntService } from '../service/defunt.service';
+import { MatDialog } from '@angular/material';
+import { DialogDeleteComponent } from '../dialog-delete/dialog-delete.component';
 
 
 @Component({
@@ -15,7 +17,7 @@ export class ListDefuntComponent implements OnInit {
    searchText: DefuntSearch;
    p = 1;
 
-  constructor(private defuntService: DefuntService) {
+  constructor(private defuntService: DefuntService, public dialog: MatDialog) {
      this.searchText = new DefuntSearch();
    }
 
@@ -26,12 +28,24 @@ export class ListDefuntComponent implements OnInit {
     });
   }
 
-  deleteDefunt (id: string) {
-    const defunt: Defunt = new Defunt(id) ;
-    this.defuntService.deleteHttp('http://localhost:3000/Defunt', defunt).subscribe((results) => {
-      this.defunts = this.defunts.filter((e) => e._id !== id);
-      console.log(results);
+  openDialog(id: string): void {
+    const dialogRef = this.dialog.open(DialogDeleteComponent, {
+      width: '300px',
+      data: { id: id }
     });
+    dialogRef.afterClosed().subscribe(result => {
+     if (result && result.id) {
+        const defunt: Defunt = new Defunt(result.id) ;
+        this.defuntService.deleteHttp('http://localhost:3000/Defunt', defunt).subscribe((results) => {
+          this.defunts = this.defunts.filter((e) => e._id !== id);
+          console.log(results);
+        });
+     }
+    });
+  }
+
+  deleteDefunt (id: string) {
+   this.openDialog(id);
   }
 
 }
